+++ projects/packages/forms/src/dashboard/inbox/index.js
@@ -1,14 +1,17 @@
 import { Gridicon } from '@automattic/jetpack-components';
 import { TabPanel } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
-import { useCallback, useEffect, useState } from '@wordpress/element';
+import { dateI18n } from '@wordpress/date';
+import { useCallback, useEffect, useMemo, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import classnames from 'classnames';
 import { find, includes, map } from 'lodash';
+import DropdownFilter from '../components/dropdown-filter';
 import Layout from '../components/layout';
 import SearchForm from '../components/search-form';
 import { STORE_NAME } from '../state';
 import BulkActionsMenu from './bulk-actions-menu';
+import ExportModal from './export-modal';
 import InboxList from './list';
 import InboxResponse from './response';
 import './style.scss';
@@ -35,16 +38,36 @@ const TABS = [
 
 const Inbox = () => {
 	const [ currentResponseId, setCurrentResponseId ] = useState( -1 );
+	const [ showExportModal, setShowExportModal ] = useState( false );
 	const [ view, setView ] = useState( 'list' );
 
-	const { fetchResponses, setCurrentPage, setSearchQuery, setStatusQuery } =
-		useDispatch( STORE_NAME );
-	const [ currentPage, loading, responses, query, total ] = useSelect(
+	const {
+		fetchResponses,
+		setCurrentPage,
+		setMonthQuery,
+		setSearchQuery,
+		setSourceQuery,
+		setStatusQuery,
+		selectResponses,
+	} = useDispatch( STORE_NAME );
+	const [
+		currentPage,
+		monthFilter,
+		sourceFilter,
+		loading,
+		query,
+		responses,
+		selectedResponses,
+		total,
+	] = useSelect(
 		select => [
 			select( STORE_NAME ).getCurrentPage(),
+			select( STORE_NAME ).getMonthFilter(),
+			select( STORE_NAME ).getSourceFilter(),
 			select( STORE_NAME ).isFetchingResponses(),
-			select( STORE_NAME ).getResponses(),
 			select( STORE_NAME ).getResponsesQuery(),
+			select( STORE_NAME ).getResponses(),
+			select( STORE_NAME ).getSelectedResponseIds(),
 			select( STORE_NAME ).getTotalResponses(),
 		],
 		[]
@@ -76,6 +99,49 @@ const Inbox = () => {
 		setView( 'list' );
 	}, [] );
 
+	const toggleExportModal = useCallback(
+		() => setShowExportModal( ! showExportModal ),
+		[ showExportModal, setShowExportModal ]
+	);
+
+	const monthList = useMemo( () => {
+		const list = map( monthFilter, item => {
+			const date = new Date();
+			date.setDate( 1 );
+			date.setMonth( item.month - 1 );
+
+			return {
+				label: `${ dateI18n( 'F', date ) } ${ item.year }`,
+				value: `${ item.year }${ String( item.month ).padStart( 2, '0' ) }`,
+			};
+		} );
+
+		return [
+			{
+				label: __( 'All dates', 'jetpack-forms' ),
+				value: null,
+			},
+			...list,
+		];
+	}, [ monthFilter ] );
+
+	const sourceList = useMemo( () => {
+		const list = map( sourceFilter, item => ( {
+			label: item.title,
+			value: item.id,
+		} ) );
+
+		return [
+			{
+				label: __( 'All sources', 'jetpack-forms' ),
+				value: null,
+			},
+			...list,
+		];
+	}, [ sourceFilter ] );
+
+	const showBulkActionsMenu = !! selectedResponses.length && ! loading;
+
 	const classes = classnames( 'jp-forms__inbox', {
 		'is-response-view': view === 'response',
 	} );
@@ -102,23 +168,49 @@ const Inbox = () => {
 				{ () => (
 					<>
 						<div className="jp-forms__inbox-actions">
-							<SearchForm
-								onSearch={ setSearchQuery }
-								initialValue={ query.search }
-								loading={ loading }
-							/>
-							<BulkActionsMenu />
+							{ ! showBulkActionsMenu && (
+								<>
+									<SearchForm
+										onSearch={ setSearchQuery }
+										initialValue={ query.search }
+										loading={ loading }
+									/>
+									<DropdownFilter
+										options={ monthList }
+										onChange={ setMonthQuery }
+										value={ query.month }
+									/>
+									<DropdownFilter
+										options={ sourceList }
+										onChange={ setSourceQuery }
+										value={ query.parent_id }
+									/>
+								</>
+							) }
+							{ showBulkActionsMenu && (
+								<BulkActionsMenu
+									currentView={ query.status }
+									selectedResponses={ selectedResponses }
+									setSelectedResponses={ selectResponses }
+								/>
+							) }
+
+							<button className="button button-primary export-button" onClick={ toggleExportModal }>
+								{ __( 'Export', 'jetpack-forms' ) }
+							</button>
 						</div>
 						<div className="jp-forms__inbox-content">
 							<div className="jp-forms__inbox-content-column">
 								<InboxList
+									currentPage={ currentPage }
 									currentResponseId={ currentResponseId }
 									loading={ loading }
-									setCurrentResponseId={ selectResponse }
-									responses={ responses }
-									currentPage={ currentPage }
-									setCurrentPage={ setCurrentPage }
 									pages={ Math.ceil( total / RESPONSES_FETCH_LIMIT ) }
+									responses={ responses }
+									selectedResponses={ selectedResponses }
+									setCurrentPage={ setCurrentPage }
+									setCurrentResponseId={ selectResponse }
+									setSelectedResponses={ selectResponses }
 								/>
 							</div>
 
@@ -132,6 +224,8 @@ const Inbox = () => {
 					</>
 				) }
 			</TabPanel>
+
+			<ExportModal isVisible={ showExportModal } onClose={ toggleExportModal } />
 		</Layout>
 	);
 };