+++ projects/packages/forms/src/dashboard/state/actions.js
@@ -8,9 +8,13 @@ import {
 	RESPONSES_FETCH,
 	RESPONSES_FETCH_FAIL,
 	RESPONSES_FETCH_RECEIVE,
+	RESPONSES_LOADING_SET,
+	RESPONSES_QUERY_MONTH_UPDATE,
 	RESPONSES_QUERY_RESET,
 	RESPONSES_QUERY_SEARCH_UPDATE,
+	RESPONSES_QUERY_SOURCE_UPDATE,
 	RESPONSES_QUERY_STATUS_UPDATE,
+	RESPONSES_SELECTION_SET,
 } from './action-types';
 
 /**
@@ -43,6 +47,7 @@ export function* fetchResponses( query ) {
 			type: RESPONSES_FETCH_RECEIVE,
 			responses: data.responses,
 			total: data.totals[ query.status || 'inbox' ],
+			filters: data.filters_available,
 		};
 	} catch ( error ) {
 		return {
@@ -73,6 +78,17 @@ export const resetQuery = () => ( {
 } );
 
 /**
+ * Update the month filter in the current query.
+ *
+ * @param {string} month - Month filter.
+ * @returns {object} Action object.
+ */
+export const setMonthQuery = month => ( {
+	type: RESPONSES_QUERY_MONTH_UPDATE,
+	month,
+} );
+
+/**
  * Update the search term in the current query.
  *
  * @param {string} search - Search term.
@@ -84,6 +100,17 @@ export const setSearchQuery = search => ( {
 } );
 
 /**
+ * Update the source filter in the current query.
+ *
+ * @param {string} source - Source filter.
+ * @returns {object} Action object.
+ */
+export const setSourceQuery = source => ( {
+	type: RESPONSES_QUERY_SOURCE_UPDATE,
+	source,
+} );
+
+/**
  * Update the status in the current query.
  *
  * @param {string} status - Feedback status.
@@ -92,4 +119,26 @@ export const setSearchQuery = search => ( {
 export const setStatusQuery = status => ( {
 	type: RESPONSES_QUERY_STATUS_UPDATE,
 	status,
+} );
+
+/**
+ * Updates the currently selected responses.
+ *
+ * @param  {Array} selectedResponses - Selected responses.
+ * @returns {object}                   Action object.
+ */
+export const selectResponses = selectedResponses => ( {
+	type: RESPONSES_SELECTION_SET,
+	selectedResponses,
+} );
+
+/**
+ * Set the application loading state.
+ *
+ * @param {boolean} loading - The loading state.
+ * @returns {object} Action object.
+ */
+export const setLoading = loading => ( {
+	type: RESPONSES_LOADING_SET,
+	loading,
 } );