+++ projects/packages/videopress/src/client/block-editor/blocks/video/edit.native.tsx
@@ -3,13 +3,14 @@
  */
 import {
 	BlockControls,
+	BlockCaption,
 	InspectorControls,
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
 import { createBlock } from '@wordpress/blocks';
 import { PanelBody } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
-import { useEffect, useState, useCallback } from '@wordpress/element';
+import { useState, useCallback, useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { store as noticesStore } from '@wordpress/notices';
 /**
@@ -20,9 +21,9 @@ import { View } from 'react-native';
 /**
  * Internal dependencies
  */
-import getMediaToken from '../../../lib/get-media-token/index.native';
 import { buildVideoPressURL, getVideoPressUrl } from '../../../lib/url';
 import { usePreview } from '../../hooks/use-preview';
+import isLocalFile from '../../utils/is-local-file.native';
 import ColorPanel from './components/color-panel';
 import DetailsPanel from './components/details-panel';
 import PlaybackPanel from './components/playback-panel';
@@ -42,6 +43,7 @@ import type { VideoBlockAttributes } from './types';
  * @param {Function} props.setAttributes - Function to set block attributes.
  * @param {boolean} props.isSelected	 - Whether block is selected.
  * @param {Function} props.onFocus       - Callback to notify when block should gain focus.
+ * @param {Function} props.insertBlocksAfter - Function to insert a new block after the current block.
  * @returns {React.ReactNode}            - React component.
  */
 export default function VideoPressEdit( {
@@ -50,9 +52,9 @@ export default function VideoPressEdit( {
 	setAttributes,
 	isSelected,
 	onFocus,
+	insertBlocksAfter,
 } ): React.ReactNode {
 	const {
-		autoplay,
 		controls,
 		guid,
 		loop,
@@ -74,21 +76,8 @@ export default function VideoPressEdit( {
 		isReplacing: false,
 		prevAttrs: {},
 	} );
-	const [ , setToken ] = useState< string >();
 
-	// Fetch token for a VideoPress GUID
-	useEffect( () => {
-		if ( guid ) {
-			getMediaToken( 'playback', { guid } )
-				.then( tokenData => {
-					setToken( tokenData.token );
-				} )
-				.catch( error => {
-					// eslint-disable-next-line no-console
-					console.error( "Can't obtain the token:", error );
-				} );
-		}
-	}, [ guid ] );
+	const [ showReplaceControl, setShowReplaceControl ] = useState( true );
 
 	const wasBlockJustInserted = useSelect(
 		select => select( blockEditorStore ).wasBlockJustInserted( clientId, 'inserter_menu' ),
@@ -98,7 +87,7 @@ export default function VideoPressEdit( {
 	const { createErrorNotice } = useDispatch( noticesStore );
 
 	const videoPressUrl = getVideoPressUrl( guid, {
-		autoplay,
+		autoplay: false, // Note: Autoplay is disabled to prevent the video from playing fullscreen when loading the editor.
 		controls,
 		loop,
 		muted,
@@ -112,10 +101,21 @@ export default function VideoPressEdit( {
 
 	const { preview, isRequestingEmbedPreview } = usePreview( videoPressUrl );
 
+	// Display upload progress in case the editor is closed and re-opened
+	// while the upload is in progress.
+	useEffect( () => {
+		const { id, src } = attributes;
+		const isUploadInProgress = !! id && ! guid && isLocalFile( src );
+		if ( isUploadInProgress ) {
+			setIsUploadingFile( true );
+			setFileToUpload( { id, url: src } );
+		}
+	}, [] );
+
 	// Handlers of `VideoPressUploader`
 	const onStartUpload = useCallback(
 		media => {
-			setAttributes( { id: media.id } );
+			setAttributes( { id: media.id, src: media.url } );
 		},
 		[ setAttributes ]
 	);
@@ -131,12 +131,13 @@ export default function VideoPressEdit( {
 
 				// Delete attributes that are not needed.
 				delete newBlockAttributes.poster;
+				delete newBlockAttributes.src;
 
 				setIsReplacingFile( { isReplacing: false, prevAttrs: {} } );
 				replaceBlock( clientId, createBlock( 'videopress/video', newBlockAttributes ) );
 				return;
 			}
-			setAttributes( { id: newVideoData.id, guid: newVideoData.guid } );
+			setAttributes( { id: newVideoData.id, guid: newVideoData.guid, src: undefined } );
 		},
 		[ setIsUploadingFile, setAttributes ]
 	);
@@ -152,6 +153,7 @@ export default function VideoPressEdit( {
 		media => {
 			setIsReplacingFile( { isReplacing: true, prevAttrs: attributes } );
 			setIsUploadingFile( true );
+			setAttributes( { guid: null } );
 			setFileToUpload( media );
 		},
 		[ setIsReplacingFile, setIsUploadingFile, setFileToUpload ]
@@ -187,6 +189,26 @@ export default function VideoPressEdit( {
 		[ setAttributes ]
 	);
 
+	const accessibilityLabelCreator = useCallback( caption => {
+		if ( caption ) {
+			return sprintf(
+				/* translators: accessibility text. %s: Video caption. */
+				__( 'Video caption. %s', 'jetpack-videopress-pkg' ),
+				caption
+			);
+		}
+		/* translators: accessibility text. Empty Video caption. */
+		return __( 'Video caption. Empty', 'jetpack-videopress-pkg' );
+	}, [] );
+
+	const onCaptionFocus = useCallback( () => {
+		setShowReplaceControl( false );
+	}, [ setShowReplaceControl ] );
+
+	const onCaptionBlur = useCallback( () => {
+		setShowReplaceControl( true );
+	}, [ setShowReplaceControl ] );
+
 	if ( isUploadingFile ) {
 		return (
 			<VideoPressUploader
@@ -205,11 +227,13 @@ export default function VideoPressEdit( {
 	return (
 		<View style={ style[ 'wp-block-jetpack-videopress__container' ] }>
 			<BlockControls>
-				<ReplaceControl
-					onUploadFileStart={ onReplaceUploadStart }
-					onSelectVideoFromLibrary={ onReplaceSelectFromLibrary }
-					onSelectURL={ onReplaceSelectURL }
-				/>
+				{ showReplaceControl && (
+					<ReplaceControl
+						onUploadFileStart={ onReplaceUploadStart }
+						onSelectVideoFromLibrary={ onReplaceSelectFromLibrary }
+						onSelectURL={ onReplaceSelectURL }
+					/>
+				) }
 			</BlockControls>
 
 			{ isSelected && (
@@ -227,6 +251,15 @@ export default function VideoPressEdit( {
 				html={ preview.html }
 				isRequestingEmbedPreview={ isRequestingEmbedPreview }
 				isSelected={ isSelected }
+			/>
+
+			<BlockCaption
+				clientId={ clientId }
+				insertBlocksAfter={ insertBlocksAfter }
+				accessibilityLabelCreator={ accessibilityLabelCreator }
+				accessible
+				onFocus={ onCaptionFocus }
+				onBlur={ onCaptionBlur }
 			/>
 		</View>
 	);