+++ projects/packages/forms/CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2023-04-04
+### Added
+- Export modal for the new JP Forms dashboard. [#29775]
+- Forms: Add single actions menu to the Dashboard inbox view [#29848]
+- Forms: Create response inbox filters [#29694]
+
+### Changed
+- Disregard post_status when calculating available filters for form responses. [#29817]
+- Forms: Dashboard finetunings [#29789]
+- Forms: Include bulk actions menu [#29766]
+- Forms: Update Dashboard inbox columns responsiveness and sticky items style [#29914]
+- Updated form responses endpoint to embed available filter data. [#29805]
+- Updated package dependencies. [#29854]
+- Updated package dependencies. [#29857]
+
+### Fixed
+- Made feedback bulk actions more explicit and easier to work with. [#29884]
+
+## [0.10.2] - 2023-04-03
+### Changed
+- Internal updates.
+
 ## [0.10.1] - 2023-03-28
 ### Changed
 - Minor internal updates.
@@ -130,6 +152,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a new jetpack/forms package [#28409]
 - Added a public load_contact_form method for initializing the contact form module. [#28416]
 
+[0.11.0]: https://github.com/automattic/jetpack-forms/compare/v0.10.2...v0.11.0
+[0.10.2]: https://github.com/automattic/jetpack-forms/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/automattic/jetpack-forms/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/automattic/jetpack-forms/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/automattic/jetpack-forms/compare/v0.8.0...v0.9.0